+++ src/Joomlatools/Console/Command/Database/Install.php
@@ -75,6 +75,8 @@ class Install extends AbstractDatabase
     {
         parent::execute($input, $output);
 
+        $password = empty($this->mysql->password) ? '' : sprintf("-p'%s'", $this->mysql->password);
+
         $this->drop        = $input->getOption('drop');
         $this->skip_check  = $input->getOption('skip-exists-check');
 
@@ -100,13 +102,38 @@ class Install extends AbstractDatabase
 
             file_put_contents($tmp, $contents);
 
-            $password = empty($this->mysql->password) ? '' : sprintf("-p'%s'", $this->mysql->password);
             $result = exec(sprintf("mysql --host=%s --port=%u --user='%s' %s %s < %s", $this->mysql->host, $this->mysql->port, $this->mysql->user, $password, $this->target_db, $tmp));
 
             unlink($tmp);
 
             if (!empty($result)) {
                 throw new \RuntimeException(sprintf('Cannot import database "%s". Error: %s', basename($import), $result));
+            }
+        }
+
+        // Fix the #__schemas table for Joomla CMS 2.5+
+        $path = $this->target_dir . '/administrator/components/com_admin/sql/updates/mysql/';
+        if (is_dir($path))
+        {
+            $updates = glob("$path/*.sql");
+
+            if (count($updates))
+            {
+                natsort($updates);
+
+                $schema = substr(basename(array_pop($updates)), 0, -4);
+                $schema = preg_replace('/[^a-z0-9\.\-\_]/i', '', $schema);
+
+                $version = Util::getJoomlaVersion($this->target_dir);
+
+                $executeQuery = function ($sql) use ($password) {
+                    $command = sprintf("mysql --host=%s --port=%u --user='%s' %s %s -e %s", $this->mysql->host, $this->mysql->port, $this->mysql->user, $password, $this->target_db, escapeshellarg($sql));
+
+                    exec ($command);
+                };
+
+                $executeQuery("REPLACE INTO j_schemas (extension_id, version_id) VALUES (700, '$schema');");
+                $executeQuery("UPDATE j_extensions SET manifest_cache = '{\"version\": \"$version->release\"}' WHERE manifest_cache = '';");
             }
         }
     }
@@ -135,10 +162,10 @@ class Install extends AbstractDatabase
 
             $version = Util::getJoomlaVersion($this->target_dir);
 
-            if($version)
+            if($version !== false && $version->release)
             {
-                if (in_array($sample_data, array('testing', 'learn')) && version_compare($version, '3.0.0', '<')) {
-                    throw new \RuntimeException(sprintf('%s does not support sample data %s', $version, $sample_data));
+                if (in_array($sample_data, array('testing', 'learn')) && version_compare($version->release, '3.0.0', '<')) {
+                    throw new \RuntimeException(sprintf('%s does not support sample data %s', $version->release, $sample_data));
                 }
             }
         }
@@ -163,14 +190,17 @@ class Install extends AbstractDatabase
         $version = Util::getJoomlaVersion($this->target_dir);
         $imports = $this->_getInstallFiles($input->getOption('sample-data'));
 
-        if ($version !== false && !Util::isPlatform($this->target_dir))
+        $isJoomlaCMS = !Util::isPlatform($this->target_dir) && !Util::isKodekitPlatform($this->target_dir);
+
+        if ($version !== false && $isJoomlaCMS)
         {
             $users = 'joomla3.users.sql';
-            if(is_numeric(substr($version, 0, 1)) && version_compare($version, '3.0.0', '<')) {
+            if(is_numeric(substr($version->release, 0, 1)) && version_compare($version->release, '3.0.0', '<')) {
                 $users = 'joomla2.users.sql';
             }
+            $path = Util::getTemplatePath();
 
-            $imports[] = self::$files.'/'.$users;
+            $imports[] = $path.'/'.$users;
         }
 
         foreach ($imports as $import)
@@ -187,7 +217,7 @@ class Install extends AbstractDatabase
     {
         $files = array();
 
-        if (Util::isPlatform($this->target_dir))
+        if (Util::isPlatform($this->target_dir) || Util::isKodekitPlatform($this->target_dir))
         {
             $path = $this->target_dir .'/install/mysql/';